refactor: refactor component type define and separate type
