+++ src/api/pr/pr.ts
@@ -1,20 +1,18 @@
+import { githubHttps } from "@/lib/github";
 import {
   GetFileDataParams,
   GetPrDataParams,
-  GitHubFileChange,
+  GitHubFileChangeResponse,
   GitHubPrResponse,
-} from "@/types/pr";
-import axios from "axios";
-
-const PR_BASE_URL = import.meta.env.VITE_PR_URL;
+} from "./dto";
 
 export const getPrData = async ({
   owner,
   repo,
   prNumber,
 }: GetPrDataParams): Promise<GitHubPrResponse> => {
-  const response = await axios.get(
-    `${PR_BASE_URL}/repos/${owner}/${repo}/pulls/${prNumber}`,
+  const response = await githubHttps.get(
+    `/repos/${owner}/${repo}/pulls/${prNumber}`,
   );
   return response.data;
 };
@@ -23,9 +21,9 @@ export const getPrCommitsData = async ({
   owner,
   repo,
   prNumber,
-}: GetPrDataParams): Promise<GitHubFileChange[]> => {
-  const response = await axios.get(
-    `${PR_BASE_URL}/repos/${owner}/${repo}/pulls/${prNumber}/files`,
+}: GetPrDataParams): Promise<GitHubFileChangeResponse[]> => {
+  const response = await githubHttps.get(
+    `/repos/${owner}/${repo}/pulls/${prNumber}/files`,
   );
   return response.data;
 };
@@ -35,9 +33,9 @@ export const getFileData = async ({
   repo,
   branchName,
   fileName,
-}: GetFileDataParams): Promise<any> => {
-  const response = await axios.get(
-    `${PR_BASE_URL}/raw/${owner}/${repo}/${branchName}/${fileName}`,
+}: GetFileDataParams) => {
+  const response = await githubHttps.get(
+    `/raw/${owner}/${repo}/${branchName}/${fileName}`,
   );
   return response.data;
 };
